+++ src/imp/order_kwargs.rs
@@ -1,5 +1,5 @@
 use crate::{
-  fun::{Book, Name},
+  fun::{parser::ParseBook, Name},
   imp::{Definition, Expr, Stmt},
 };
 use indexmap::IndexMap;
@@ -7,92 +7,99 @@ use indexmap::IndexMap;
 impl Definition {
   /// Traverses the program's definitions and adjusts the order of keyword arguments
   /// in call/constructor expressions to match the order specified in the function or constructor definition.
-  pub fn order_kwargs(&mut self, book: &Book) -> Result<(), String> {
-    self.body.order_kwargs(book).map_err(|e| format!("In function '{}':\n  {}", self.name, e))
+  pub fn order_kwargs(&mut self, book: &ParseBook) -> Result<(), String> {
+    let use_map = &mut IndexMap::new();
+    self.body.order_kwargs(book, use_map).map_err(|e| format!("In function '{}':\n  {}", self.name, e))
   }
 }
 
 impl Stmt {
-  fn order_kwargs(&mut self, book: &Book) -> Result<(), String> {
+  fn order_kwargs(&mut self, book: &ParseBook, use_map: &mut IndexMap<Name, Name>) -> Result<(), String> {
     match self {
       Stmt::LocalDef { def, nxt } => {
         def.order_kwargs(book)?;
-        nxt.order_kwargs(book)?;
+        nxt.order_kwargs(book, use_map)?;
       }
       Stmt::Assign { val, nxt, .. } => {
-        val.order_kwargs(book)?;
+        val.order_kwargs(book, use_map)?;
         if let Some(nxt) = nxt {
-          nxt.order_kwargs(book)?;
+          nxt.order_kwargs(book, use_map)?;
         }
       }
       Stmt::Ask { val, nxt, .. } => {
-        val.order_kwargs(book)?;
-        nxt.order_kwargs(book)?;
+        val.order_kwargs(book, use_map)?;
+        nxt.order_kwargs(book, use_map)?;
       }
       Stmt::InPlace { val, nxt, .. } => {
-        val.order_kwargs(book)?;
-        nxt.order_kwargs(book)?;
+        val.order_kwargs(book, use_map)?;
+        nxt.order_kwargs(book, use_map)?;
       }
       Stmt::If { cond, then, otherwise, nxt } => {
-        cond.order_kwargs(book)?;
-        then.order_kwargs(book)?;
-        otherwise.order_kwargs(book)?;
+        cond.order_kwargs(book, use_map)?;
+        then.order_kwargs(book, use_map)?;
+        otherwise.order_kwargs(book, use_map)?;
         if let Some(nxt) = nxt {
-          nxt.order_kwargs(book)?;
+          nxt.order_kwargs(book, use_map)?;
         }
       }
       Stmt::Match { arg, arms, nxt, .. } => {
-        arg.order_kwargs(book)?;
+        arg.order_kwargs(book, use_map)?;
         for arm in arms {
-          arm.rgt.order_kwargs(book)?;
+          arm.rgt.order_kwargs(book, use_map)?;
         }
         if let Some(nxt) = nxt {
-          nxt.order_kwargs(book)?;
+          nxt.order_kwargs(book, use_map)?;
         }
       }
       Stmt::Switch { arg, arms, nxt, .. } => {
-        arg.order_kwargs(book)?;
+        arg.order_kwargs(book, use_map)?;
         for arm in arms {
-          arm.order_kwargs(book)?;
+          arm.order_kwargs(book, use_map)?;
         }
         if let Some(nxt) = nxt {
-          nxt.order_kwargs(book)?;
+          nxt.order_kwargs(book, use_map)?;
         }
       }
       Stmt::Fold { arg, arms, nxt, .. } => {
-        arg.order_kwargs(book)?;
+        arg.order_kwargs(book, use_map)?;
         for arm in arms {
-          arm.rgt.order_kwargs(book)?;
+          arm.rgt.order_kwargs(book, use_map)?;
         }
         if let Some(nxt) = nxt {
-          nxt.order_kwargs(book)?;
+          nxt.order_kwargs(book, use_map)?;
         }
       }
       Stmt::Bend { bnd: _, arg, cond, step, base, nxt } => {
         for arg in arg {
-          arg.order_kwargs(book)?;
+          arg.order_kwargs(book, use_map)?;
         }
-        cond.order_kwargs(book)?;
-        step.order_kwargs(book)?;
-        base.order_kwargs(book)?;
+        cond.order_kwargs(book, use_map)?;
+        step.order_kwargs(book, use_map)?;
+        base.order_kwargs(book, use_map)?;
         if let Some(nxt) = nxt {
-          nxt.order_kwargs(book)?;
+          nxt.order_kwargs(book, use_map)?;
         }
       }
       Stmt::With { typ: _, bod, nxt } => {
-        bod.order_kwargs(book)?;
+        bod.order_kwargs(book, use_map)?;
         if let Some(nxt) = nxt {
-          nxt.order_kwargs(book)?;
+          nxt.order_kwargs(book, use_map)?;
         }
       }
       Stmt::Open { typ: _, var: _, nxt } => {
-        nxt.order_kwargs(book)?;
+        nxt.order_kwargs(book, use_map)?;
       }
-      Stmt::Use { nam: _, val: bod, nxt } => {
-        bod.order_kwargs(book)?;
-        nxt.order_kwargs(book)?;
+      Stmt::Use { nam, val: bod, nxt } => {
+        if let Expr::Var { nam: bod } = bod.as_ref() {
+          use_map.insert(nam.clone(), bod.clone());
+          nxt.order_kwargs(book, use_map)?;
+          use_map.pop();
+        } else {
+          bod.order_kwargs(book, use_map)?;
+          nxt.order_kwargs(book, use_map)?;
+        }
       }
-      Stmt::Return { term } => term.order_kwargs(book)?,
+      Stmt::Return { term } => term.order_kwargs(book, use_map)?,
       Stmt::Err => {}
     }
     Ok(())
@@ -100,13 +107,13 @@ impl Stmt {
 }
 
 impl Expr {
-  fn order_kwargs(&mut self, book: &Book) -> Result<(), String> {
+  fn order_kwargs(&mut self, book: &ParseBook, use_map: &mut IndexMap<Name, Name>) -> Result<(), String> {
     match self {
       // Named arguments are only allowed when directly calling a named function.
       Expr::Call { fun, args, kwargs } => {
         if !kwargs.is_empty() {
           if let Expr::Var { nam } = fun.as_ref() {
-            if let Some(names) = get_args_def_or_ctr(nam, book) {
+            if let Some(names) = get_args_def_or_ctr(nam, book, use_map) {
               go_order_kwargs(&names, args, kwargs)?;
             } else {
               return Err(format!(
@@ -121,54 +128,54 @@ impl Expr {
             );
           }
         }
-        fun.order_kwargs(book)?;
+        fun.order_kwargs(book, use_map)?;
         for arg in args {
-          arg.order_kwargs(book)?;
+          arg.order_kwargs(book, use_map)?;
         }
         for (_, arg) in kwargs {
-          arg.order_kwargs(book)?;
+          arg.order_kwargs(book, use_map)?;
         }
       }
-      Expr::Lam { bod, .. } => bod.order_kwargs(book)?,
+      Expr::Lam { bod, .. } => bod.order_kwargs(book, use_map)?,
       Expr::Opr { lhs, rhs, .. } => {
-        lhs.order_kwargs(book)?;
-        rhs.order_kwargs(book)?;
+        lhs.order_kwargs(book, use_map)?;
+        rhs.order_kwargs(book, use_map)?;
       }
       Expr::Lst { els } | Expr::Tup { els } | Expr::Sup { els } => {
         for el in els {
-          el.order_kwargs(book)?;
+          el.order_kwargs(book, use_map)?;
         }
       }
       Expr::LstMap { term, iter, cond, .. } => {
-        term.order_kwargs(book)?;
-        iter.order_kwargs(book)?;
+        term.order_kwargs(book, use_map)?;
+        iter.order_kwargs(book, use_map)?;
         if let Some(cond) = cond {
-          cond.order_kwargs(book)?;
+          cond.order_kwargs(book, use_map)?;
         }
       }
-      Expr::Ctr { name, args, kwargs } => match get_args_def_or_ctr(name, book) {
+      Expr::Ctr { name, args, kwargs } => match get_args_def_or_ctr(name, book, use_map) {
         Some(names) => {
           go_order_kwargs(&names, args, kwargs)?;
           for arg in args {
-            arg.order_kwargs(book)?;
+            arg.order_kwargs(book, use_map)?;
           }
         }
         _ => return Err(format!("Constructor '{name}' not found.")),
       },
       Expr::Map { entries } => {
         for entry in entries {
-          entry.1.order_kwargs(book)?;
+          entry.1.order_kwargs(book, use_map)?;
         }
       }
       Expr::MapGet { nam: _, key } => {
-        key.order_kwargs(book)?;
+        key.order_kwargs(book, use_map)?;
       }
       Expr::TreeNode { left, right } => {
-        left.order_kwargs(book)?;
-        right.order_kwargs(book)?;
+        left.order_kwargs(book, use_map)?;
+        right.order_kwargs(book, use_map)?;
       }
       Expr::TreeLeaf { val } => {
-        val.order_kwargs(book)?;
+        val.order_kwargs(book, use_map)?;
       }
       Expr::Era | Expr::Var { .. } | Expr::Chn { .. } | Expr::Num { .. } | Expr::Str { .. } => {}
     }
@@ -201,11 +208,13 @@ fn go_order_kwargs(
   Ok(())
 }
 
-fn get_args_def_or_ctr(name: &Name, book: &Book) -> Option<Vec<Name>> {
+fn get_args_def_or_ctr(name: &Name, book: &ParseBook, use_map: &IndexMap<Name, Name>) -> Option<Vec<Name>> {
+  let name = use_map.get(name).unwrap_or(name);
+
   #[allow(clippy::manual_map)]
   if let Some(adt_nam) = book.ctrs.get(name) {
     Some(book.adts[adt_nam].ctrs[name].iter().map(|f| f.nam.clone()).collect())
-  } else if let Some(def) = book.defs.get(name) {
+  } else if let Some(def) = book.fun_defs.get(name) {
     Some(def.rules[0].pats.iter().flat_map(|p| p.binds().flatten().cloned()).collect())
   } else {
     None