+++ src/imp/to_fun.rs
@@ -2,11 +2,30 @@ use super::{AssignPattern, Definition, Expr, InPlaceOp, Stmt};
 use crate::fun::{
   self,
   builtins::{LCONS, LNIL},
-  Name,
+  parser::ParseBook,
+  Book, Name,
 };
 
+impl ParseBook {
+  pub fn to_fun(mut self) -> Result<Book, String> {
+    for (name, mut def) in std::mem::take(&mut self.imp_defs) {
+      def.order_kwargs(&self)?;
+      def.gen_map_get();
+
+      if self.fun_defs.contains_key(&name) {
+        panic!("Def names collision should be checked at parse time")
+      }
+
+      self.fun_defs.insert(name, def.to_fun()?);
+    }
+
+    let ParseBook { fun_defs: defs, hvm_defs, adts, ctrs, import_ctx, .. } = self;
+    Ok(Book { defs, hvm_defs, adts, ctrs, entrypoint: None, imports: import_ctx.to_imports() })
+  }
+}
+
 impl Definition {
-  pub fn to_fun(self, builtin: bool) -> Result<fun::Definition, String> {
+  pub fn to_fun(self) -> Result<fun::Definition, String> {
     let body = self.body.into_fun().map_err(|e| format!("In function '{}': {}", self.name, e))?;
     let body = match body {
       StmtToFun::Return(term) => term,
@@ -18,7 +37,7 @@ impl Definition {
     let rule =
       fun::Rule { pats: self.params.into_iter().map(|param| fun::Pattern::Var(Some(param))).collect(), body };
 
-    let def = fun::Definition { name: self.name, rules: vec![rule], builtin };
+    let def = fun::Definition::new(self.name, vec![rule], self.source);
     Ok(def)
   }
 }
@@ -371,7 +390,7 @@ impl Stmt {
           StmtToFun::Return(term) => (None, term),
           StmtToFun::Assign(pat, term) => (Some(pat), term),
         };
-        let def = def.to_fun(false)?;
+        let def = def.to_fun()?;
         let term = fun::Term::Def { nam: def.name, rules: def.rules, nxt: Box::new(nxt) };
         if let Some(pat) = nxt_pat {
           StmtToFun::Assign(pat, term)
@@ -420,7 +439,7 @@ impl Expr {
       Expr::Ctr { name, args, kwargs } => {
         assert!(kwargs.is_empty());
         let args = args.into_iter().map(Self::to_fun);
-        fun::Term::call(fun::Term::Ref { nam: name }, args)
+        fun::Term::call(fun::Term::Var { nam: name }, args)
       }
       Expr::LstMap { term, bind, iter, cond } => {
         const ITER_TAIL: &str = "%iter.tail";